Added scenario calculations
Trying to re run life tables with modfified incidence. I added three columns to each of the disease life tables (list) for the difference in incidence, mortatliy and prevalence.  My code is not re running the life tables with the mofified incidence as there are no differences in incidnece, prevalence and mortality